Removed transparent mode
Removed as it was not effective. It was intended to enable entering K197 self-test mode. Probably the only way to enter test mode would require a HW change. This may be added if and when we revise the HW. A workaround would be to power the display board independently, hold the Db button and then power up the voltmeter.