Moved to manual interrupt handler for I/O pins
Since on rare occasion SPIdevice goes out of symnch, we have moved to a more efficient way to handle the I/O pin interrupts available with with dxCore. Instead of using attachInterrupt (which is very innefficiant) we can now define the interrupt vecrtor with the ISR() macro.
In addition, we have done further optimizations replacing a couple of variables using in the interrupt handler with GPIOR1 and GPIOR2 which were free to use