Bug fix in class MenuInputByte (increment/decrement in UI)
Fixed the problem with incrementing/decrementing when setValue and getValue are overridden in a child class. Also improved performance since now setValue is only invoked at button release. A new boolean member in the class was needed to keep track of the fact we are editing the value in the UI