class K197device: restructured code caching main unit and prefix
The change is needed because the previous method relied on extracting the first and second char from the string returned from getUnit. This would not work because getUnit return an unicode string. To avoid the complication with multibyte characters, we have implemented a new function getMainUnit() and for the prefix we rely on getUnitPow10()