Potential bug fix
we need to protect access to nbyte and spiflags since they are used inside the interrupt handler. Normally this should not have lead to problems because we read and modify them as soon as all SPI data is received, and it should take 0.3 s before the next data is received from the SPI intercface. But it's better to be sure and eliminate a potential cauise of problems