class K197PushButtons: interfaced to Ev system and CCL directly
The Event and Logic libraries have been useful to develop the SW, however now that the implementation of this part is stable interfacinjg direcrtly to the regoisters saves about 1K flash abd 170 bytes of RAM, both needed to continue the development/troubleshooting. Furthermore, we remove a lot of compilation warnings generated by the Logic library.