Restrucutred graph code (both storage and display)
The way the graph information was handled was too complex. this is due to the fact we implemented the so called overwroite mode and then we added roll mode on top. Now that overwrite mode is not used anymore, we can restructre the code as we don't need anymore to distinguish between raay index and logical index outside the class that stores the graph.