Minor improvement: added some protection to the graph from some measurement  glitches
In rare occasions there are measurement glitches due to various reasons. When we detect a change in measurement conditions (unit, etc.), rather than resetting the graph immediately we wait for the next sample. If the change is confirmed, then we reset the graph. Otherwise we continue, possibly skipping a sample (the one with the glitch).